pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "analemma-capture"
version = "1.0.0"
description = "Automated solar photography system for capturing analemma patterns"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Analemma Project" }
]
keywords = ["astronomy", "solar", "analemma", "photography", "zwo", "raspberry-pi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
]
dependencies = [
    "zwoasi>=0.1.0",
    "apscheduler>=3.10.0",
    "astropy>=6.0.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
]

[project.scripts]
analemma = "analemma.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/analemma"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
